flet_app/pyproject.toml
[project]
name = "omni-connect"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Andrii Bogdanovych", email = "ab@foundation101.org" }
]
dependencies = [
  "flet>=0.80.2",
  "flet-storage>=1.0.0",
  "pyserial>=3.5",
]

[tool.flet]
org = "org.foundation101"
product = "Omni Connect"
company = "NGO Foundation.101"
copyright = "Copyright (C) 2026 by NGO Foundation.101"

[tool.flet.app]
path = "src"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']
skip-string-normalization = true

[tool.isort]
profile = "black"